// src/types/star-wars-character.ts
export type StarWarsCharacter = {
  name: string;
  height: string;
  mass: string;
  hair_color: string;
  skin_color: string;
  eye_color: string;
  birth_year: string;
  gender: StarWarsCharacterGender;
  homeworld: string;
  films: string[];
  species: string[];
  vehicles: string[];
  starships: string[];
  created: string;
  edited: string;
  url: string;
};

export type StarWarsCharacterGender = "male" | "female" | "n/a";